requirements.txt
requests>=2.28,<3.0
python-dotenv>=1.0,<2.0
matplotlib>=3.5,<4.0